table changes to support prices as integers

+++ backend/db/migrations/20230901055005-create-spot.js
@@ -54,7 +54,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.DECIMAL(6,2)
+        type: Sequelize.INTEGER
       },
       previewUrl: {
         allowNull: false,
